move Identifiable to ml.util
